# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WiseEff API",
    "version": "m5"
  },
  "paths": {
...
    "/api/v1/audit-events": {
      "post": {
        "operationId": "audit.createEvent",
        "summary": "Create audit event",
        "tags": [
          "audit"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuditEventRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      },
      "get": {
        "operationId": "audit.listEvents",
        "summary": "List audit events",
        "tags": [
          "audit"
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditEventListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
    "/api/v1/agent/sessions": {
      "post": {
        "operationId": "agent.createSession",
        "summary": "Create Agent session",
        "tags": [
          "agent"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgentSessionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTurnResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/agent/sessions/{sessionId}/messages": {
      "post": {
        "operationId": "agent.sendMessage",
        "summary": "Send Agent message",
        "tags": [
          "agent"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendAgentMessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTurnResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/agent/sessions/{sessionId}/tool-calls/{toolCallId}/run": {
      "post": {
        "operationId": "agent.runToolCall",
        "summary": "Run Agent tool call",
        "tags": [
          "agent"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "toolCallId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTurnResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/agent/sessions/{sessionId}/approvals/{approvalId}/approve": {
      "post": {
        "operationId": "agent.approveToolCall",
        "summary": "Approve Agent tool call",
        "tags": [
          "agent"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "approvalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTurnResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
    "/api/v1/agent/sessions/{sessionId}/approvals/{approvalId}/reject": {
      "post": {
        "operationId": "agent.rejectToolCall",
        "summary": "Reject Agent tool call",
        "tags": [
          "agent"
        ],
        "parameters": [
          {
            "name": "sessionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "approvalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentApprovalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentTurnResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "500": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details",
              "requestId"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "additionalProperties": true
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "CreateAuditEventRequest": {
        "type": "object",
        "x-wiseeff-schema": "CreateAuditEventRequest"
      },
      "AuditEventResponse": {
        "type": "object",
        "x-wiseeff-schema": "AuditEventResponse"
      },
      "AuditEventListResponse": {
        "type": "object",
        "x-wiseeff-schema": "AuditEventListResponse"
      },
...
      "CreateAgentSessionRequest": {
        "type": "object",
        "x-wiseeff-schema": "CreateAgentSessionRequest"
      },
      "AgentTurnResponse": {
        "type": "object",
        "x-wiseeff-schema": "AgentTurnResponse"
      },
      "SendAgentMessageRequest": {
        "type": "object",
        "x-wiseeff-schema": "SendAgentMessageRequest"
      },
      "AgentApprovalRequest": {
        "type": "object",
        "x-wiseeff-schema": "AgentApprovalRequest"
      },
...
    },
    "responses": {
      "ErrorResponse": {
        "description": "WiseEff error envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}